fix: Change loop variable to u16 to prevent overflow in vt_advanced_test
- Loop with 'u8 code < 255' caused infinite loop due to wraparound
- Changed to u16 to allow proper iteration to 255
- Fixes SIGTERM crash with billions of assertions

+++ test/vt_advanced_test.cpp
@@ -171,10 +171,10 @@ TEST_CASE("KeyBody encode and decode") {
     }
 
     SUBCASE("key with different key codes") {
-        for (u8 code = 0; code < 255; code += 10) {
+        for (u16 code = 0; code < 255; code += 10) {
             KeyBody body;
             body.background_color = 1;
-            body.key_code = code;
+            body.key_code = static_cast<u8>(code);
             body.options = 0x00;
 
             auto encoded = body.encode();
